requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy[asyncio]>=1.4.0
alembic>=1.7.0
python-dotenv>=0.19.0
pydantic-settings>=2.0.0
asyncpg>=0.25.0